+++ com.beatwaves.responsible/Runtime/Bdd/Keywords.cs
@@ -7,7 +7,7 @@ namespace Responsible.Bdd
 	/// Keywords for building BDD-style tests for clearer source code and state strings.
 	/// Helps in building tests with consistent style.
 	/// </summary>
-	public static class Keywords
+	public static partial class Keywords
 	{
 		/// <summary>
 		/// Starts building a new BDD-style scenario.